generate_pallet now saves to db. Reorganize.
generate_pallet saves to CSV and DB. The latter using
StrawLocation::forceAddStraw. tests/move_straws tests this safely.

Moved a bunch of stuff out of consolidate and generate and into
new consolidate_utils file. Logging handled properly.

No other changes to consolidate functionality.
